pyproject.toml
[tool.poetry]
name = "clt-forge"
version = "0.1.0"
description = "Training cross-layer-transcoder"
authors = ["Florent Draye"]
readme = "README.md"
packages = [
    { include = "clt_forge", from = "src" }
]

[tool.poetry.dependencies]
python = "3.11"
sae-lens = ">=5.9.1,<6.11.2"
pydantic = ">=2.11.3,<3.0.0"
pytest = ">=8.3.5,<9.0.0"
wandb = ">=0.19.10,<0.20.0"
torchvision = ">=0.22.0,<0.23.0"
matplotlib = ">=3.10.6,<4.0.0"
argparse = "^1.4.0"
zstd = "^1.5.7.3"
zstandard = "^0.25.0"
tqdm = "^4.67.3"
lmdb = "^1.7.5"
dash = "^4.0.0"
dash-cytoscape = "^1.0.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
explicit_package_bases = true
mypy_path = "src"
files = ["src"]
